// master-data-ui-proj/src/app/views/farmer/models/StressDetail.ts
export class StressDetail {
    commodityID: number;
    commodityName: string;
    districtID: number;
    districtName: string;
    farmerGivenImage: string;
    id: string;
    phenophaseID: number;
    plantPartID: number;
    plantPartName: string;
    phenophaseName: string;
    referenceImages: string;
    stageID: number;
    stageName: string;
    stressID: number;
    stressName: string;
    stressTypeID: number;
    stressTypeName: string;
    symptomDescription: string;
    symptomID: number;
    varietyID: number;
    varietyName: string;
    SymptomImageUrl: string;
}